vit-xray-classification/pneumonia_detection/requirements.txt
Flask==2.0.2
Pillow==8.3.2
transformers==4.30.2
torch==1.9.0+cu102
flask-cors
